third_party/WebKit/LayoutTests/svg/dom/SVGFESpecularLightingElement-svgdom-kernelUnitLength.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>
test(function() {
   var specularElement = document.createElementNS("http://www.w3.org/2000/svg", "feSpecularLighting");
   specularElement.setAttribute("kernelUnitLength", "3 4");
   assert_equals(specularElement.getAttribute('kernelUnitLength'), "3 4");
   assert_equals(specularElement.kernelUnitLengthX.baseVal, 3);
   assert_equals(specularElement.kernelUnitLengthY.baseVal, 4);
   specularElement.kernelUnitLengthX.baseVal = 5;
   specularElement.kernelUnitLengthY.baseVal = 10;
   assert_equals(specularElement.kernelUnitLengthX.baseVal, 5);
   assert_equals(specularElement.kernelUnitLengthY.baseVal, 10);
   assert_equals(specularElement.getAttribute('kernelUnitLength'), "5 10");
}, 'tests of the |kernelUnitLength| attribute of the SVGFESpecularLightingElement object');
</script>
